+++ src/mongo/mongo.py
@@ -13,6 +13,11 @@ async def setup():
     global client
     client = motor.motor_asyncio.AsyncIOMotorClient(os.getenv('MONGO_URL'))
 
+async def delete(db_name, collection_name, iden):
+    global client
+    collection = client[db_name][collection_name]
+    await collection.delete_one({"_id": iden})
+
 async def get_entire_collection(db_name, collection_name, return_one = False):
     global client
     collection = client[db_name][collection_name]
@@ -32,10 +37,27 @@ async def get_cron():
 async def get_censor():
     return await get_entire_collection("data", "censor", return_one = True)
 
+async def insert(db_name, collection_name, insert_dict):
+    global client
+    collection = client[db_name][collection_name]
+    return await collection.insert_one(insert_dict)
+
 async def update(db_name, collection_name, doc_id, update_dict):
     global client
     collection = client[db_name][collection_name]
     await collection.update_one({'_id': doc_id}, update_dict)
+
+async def update_many(db_name, collection_name, docs, update_dict):
+    global client
+    collection = client[db_name][collection_name]
+    ids = [doc.get("_id") for doc in docs]
+    await collection.update_many(
+        {"_id": {
+            "$in": ids
+            }
+        },
+        update_dict
+    )
 
 async def remove_doc(db_name, collection_name, doc_id):
     global client